Create 5.b
Tower of Hanoi

Objective:
To understand and implement the Tower of Hanoi problem using stack data structure in C.

Problem Statement:
The Tower of Hanoi is a classic problem that involves moving disks from one peg to another, subject to the following constraints:
1. Only one disk can be moved at a time.
2. Each move consists of taking the top disk from one stack and placing it on top of another
stack.
3. No disk may be placed on top of a smaller disk.
Implement the Tower of Hanoi problem using the stack data structure. You are required to implement the following functions:
void towerOfHanoi(int numDisks, char source, char auxiliary, char destination): 
A recursive function to solve the Tower of Hanoi problem. It takes the number of disks (numDisks), and characters representing the source, auxiliary, and destination pegs (source, auxiliary, destination).
void displayMove(char source, char destination): A function to display the move of disks from the source peg to the destination peg.

Instructions:
Implement the towerOfHanoi function using recursion to solve the Tower of Hanoi problem.
In the towerOfHanoi function, use stacks to represent the pegs. You can implement a stack using an array or linked list.
Implement the displayMove function to display the move of disks from one peg to another.
Test your implementation with different numbers of disks.

Sample Input/Output:
Enter the number of disks: 3
Moves to solve the Tower of Hanoi with 3 disks:
Move disk 1 from A to C
Move disk 2 from A to B
Move disk 1 from C to B
Move disk 3 from A to C
Move disk 1 from B to A
Move disk 2 from B to C
Move disk 1 from A to C